core/requirements.txt
Django==2.0.7
django-autocomplete-light==3.2.10
django-filter==2.0.0
django-taggit==0.22.2
django-wysiwyg-redactor==0.5.1
pytz==2018.5
six==1.11.0
